Generalize bot to process any message from Pub/Sub
Bot publishes any message from Pub/Sub subscription.
Added cloud functions to process and re-publish cloud build and monitoring messages.
