+++ go.mod
@@ -1,0 +1,8 @@
+module obsh.org/pubsub-keybase
+
+go 1.15
+
+require (
+	cloud.google.com/go/pubsub v1.8.2
+	github.com/keybase/go-keybase-chat-bot v0.0.0-20200505163032-5cacf52379da
+)
